Webkit: fix browser can't display italic font in WCSS page
With no knowledge of the actual SkTypeface used to render
text before render-time, the best we can do is turn fake
italic on for all text and simply ignore it at render-time
if the SkTypeface's style is italic.

Also the complex FontPlatformData hashing mechanism needed
to account for fake styles, as the versions of FontPlatformData
with fake styles enabled are otherwise indistinguishable from the
normal style version.

CRs-Fixed: 580264

Change-Id: Iaf0a621b954739345772c022190f9dbf3c2a5afa